Add basic mole game structure to controller
The game is still broken at this point but some major changes have been made.
A thread has been added that handles adding moles to the field. This logic will be removed
to its own class further on. If the game starts, the moles can be clicked but it is not consistent.